requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
apify-client>=1.4.0
requests>=2.31.0
